initial version of Fizzy_Json_Client
Fizzy_Json_Client makes it easy to talk to JSON-RPC servers. It does all the
JSON encoding and decoding for you. It also provides a convenient response
class giving easy access to all the JSON-RPC specific response data.

There also is an experimental __call implementation to make method calls
in a transparant PHP native like way. Example:

$responseObject = $client->sayHello();

calls the method 'sayHello' on the webservice. Note there is no support
yet for namespaces in this way of calling. For namespaces use the
setMethod way.

Classes:
- Fizzy_Json_Client
- Fizzy_Json_Client_Exception
- Fizzy_Json_Client_Response